+++ conftest.py
@@ -9,7 +9,7 @@ import tcod.ecs
 @pytest.fixture(autouse=True)
 def _add_world_entity(doctest_namespace: Dict[str, Any]) -> None:
     """Add world and entity objects to all doctests."""
-    world = tcod.ecs.World()
+    world = tcod.ecs.Registry()
     entity = world["entity"]
     other_entity = world["other"]
     doctest_namespace.update(