+++ tcod/ecs/__init__.py
@@ -7,12 +7,14 @@ from typing import TypeVar
 
 from tcod.ecs.constants import IsA
 from tcod.ecs.entity import Entity
-from tcod.ecs.world import World
+from tcod.ecs.registry import Registry
+from tcod.ecs.registry import Registry as World
 
 __all__ = (
     "__version__",
     "Entity",
     "IsA",
+    "Registry",
     "World",
 )